feat(qqPlot): Refactored directory structure, wrote CSS to show button
